[#051] File renamed, all dependencies checked
